pyproject.toml
[tool.poetry]
name = "tutorona"
version = "0.1.0"
description = "web app for code jam"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask-socketio = "^4.2"
pytz = "^2019.3"
psycopg2-binary = "^2.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
